+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.1.7...@kong-ui-public/analytics-config-store@1.1.8) (2025-09-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [1.1.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.1.6...@kong-ui-public/analytics-config-store@1.1.7) (2025-08-26)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store