+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.4...@kong-ui-public/analytics-metric-provider@10.1.5) (2025-09-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [10.1.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.3...@kong-ui-public/analytics-metric-provider@10.1.4) (2025-09-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [10.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.2...@kong-ui-public/analytics-metric-provider@10.1.3) (2025-09-12)