+++ packages/core/forms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.17.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.17.4...@kong-ui-public/forms@4.17.5) (2025-09-15)
+
+**Note:** Version bump only for package @kong-ui-public/forms
+
+
+
+
+
 ## [4.17.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/forms@4.17.3...@kong-ui-public/forms@4.17.4) (2025-09-12)