+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.30.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.30.3...@kong-ui-public/entities-shared@3.30.4) (2025-09-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
 ## [3.30.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.30.2...@kong-ui-public/entities-shared@3.30.3) (2025-09-12)